refactor: continue cleaning up
