+++ ethers-providers/src/lib.rs
@@ -10,21 +10,15 @@ pub use ext::*;
 mod rpc;
 pub use rpc::*;
 
+mod toolbox;
+pub use toolbox::*;
+
 pub mod utils;
 pub use utils::*;
 
 /// Errors
 pub mod errors;
 pub use errors::{MiddlewareError, ProviderError, RpcError};
-
-mod pending_transaction;
-pub use pending_transaction::PendingTransaction;
-
-mod pending_escalator;
-pub use pending_escalator::EscalatingPending;
-
-mod log_query;
-pub use log_query::{LogQuery, LogQueryError};
 
 mod stream;
 pub use futures_util::StreamExt;
@@ -34,8 +28,6 @@ pub use stream::{
 
 mod pubsub;
 pub use pubsub::{PubsubClient, SubscriptionStream};
-
-pub mod call_raw;
 
 pub mod middleware;
 pub use middleware::*;
