+++ ethers-providers/src/stream/watcher.rs
@@ -1,13 +1,12 @@
 #![allow(clippy::return_self_not_must_use)]
 
-use crate::{JsonRpcClient, Middleware, PinBoxFut, Provider, ProviderError};
-use ethers_core::types::{Transaction, TxHash, U256};
-use futures_core::{stream::Stream, Future};
-use futures_util::{stream, stream::FuturesUnordered, FutureExt, StreamExt};
+use crate::{JsonRpcClient, Middleware, PinBoxFut, Provider};
+use ethers_core::types::U256;
+use futures_core::stream::Stream;
+use futures_util::{stream, FutureExt, StreamExt};
 use pin_project::pin_project;
 use serde::{de::DeserializeOwned, Serialize};
 use std::{
-    collections::VecDeque,
     fmt::Debug,
     pin::Pin,
     task::{Context, Poll},
@@ -44,7 +43,7 @@ pub struct FilterWatcher<'a, P, R> {
     /// The filter's installed id on the ethereum node
     pub id: U256,
 
-    provider: &'a Provider<P>,
+    pub(crate) provider: &'a Provider<P>,
 
     // The polling interval
     interval: Box<dyn Stream<Item = ()> + Send + Unpin>,
@@ -121,136 +120,6 @@ where
                 }
             };
         }
-    }
-}
-
-impl<'a, P> FilterWatcher<'a, P, TxHash>
-where
-    P: JsonRpcClient,
-{
-    /// Returns a stream that yields the `Transaction`s for the transaction hashes this stream
-    /// yields.
-    ///
-    /// This internally calls `Provider::get_transaction` with every new transaction.
-    /// No more than n futures will be buffered at any point in time, and less than n may also be
-    /// buffered depending on the state of each future.
-    pub fn transactions_unordered(self, n: usize) -> TransactionStream<'a, P, Self> {
-        TransactionStream::new(self.provider, self, n)
-    }
-}
-
-/// Errors `TransactionStream` can throw
-#[derive(Debug, thiserror::Error)]
-pub enum GetTransactionError {
-    #[error("Failed to get transaction `{0}`: {1}")]
-    ProviderError(TxHash, ProviderError),
-    /// `get_transaction` resulted in a `None`
-    #[error("Transaction `{0}` not found")]
-    NotFound(TxHash),
-}
-
-impl From<GetTransactionError> for ProviderError {
-    fn from(err: GetTransactionError) -> Self {
-        match err {
-            GetTransactionError::ProviderError(_, err) => err,
-            err @ GetTransactionError::NotFound(_) => ProviderError::CustomError(err.to_string()),
-        }
-    }
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-type TransactionFut<'a> = Pin<Box<dyn Future<Output = TransactionResult> + Send + 'a>>;
-#[cfg(target_arch = "wasm32")]
-type TransactionFut<'a> = Pin<Box<dyn Future<Output = TransactionResult> + 'a>>;
-
-type TransactionResult = Result<Transaction, GetTransactionError>;
-
-/// Drains a stream of transaction hashes and yields entire `Transaction`.
-#[must_use = "streams do nothing unless polled"]
-pub struct TransactionStream<'a, P, St> {
-    /// Currently running futures pending completion.
-    pending: FuturesUnordered<TransactionFut<'a>>,
-    /// Temporary buffered transaction that get started as soon as another future finishes.
-    buffered: VecDeque<TxHash>,
-    /// The provider that gets the transaction
-    provider: &'a Provider<P>,
-    /// A stream of transaction hashes.
-    stream: St,
-    /// max allowed futures to execute at once.
-    max_concurrent: usize,
-}
-
-impl<'a, P: JsonRpcClient, St> TransactionStream<'a, P, St> {
-    /// Create a new `TransactionStream` instance
-    pub fn new(provider: &'a Provider<P>, stream: St, max_concurrent: usize) -> Self {
-        Self {
-            pending: Default::default(),
-            buffered: Default::default(),
-            provider,
-            stream,
-            max_concurrent,
-        }
-    }
-
-    /// Push a future into the set
-    fn push_tx(&mut self, tx: TxHash) {
-        let fut = self.provider.get_transaction(tx).then(move |res| match res {
-            Ok(Some(tx)) => futures_util::future::ok(tx),
-            Ok(None) => futures_util::future::err(GetTransactionError::NotFound(tx)),
-            Err(err) => futures_util::future::err(GetTransactionError::ProviderError(tx, err)),
-        });
-        self.pending.push(Box::pin(fut));
-    }
-}
-
-impl<'a, P, St> Stream for TransactionStream<'a, P, St>
-where
-    P: JsonRpcClient,
-    St: Stream<Item = TxHash> + Unpin + 'a,
-{
-    type Item = TransactionResult;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.get_mut();
-
-        // drain buffered transactions first
-        while this.pending.len() < this.max_concurrent {
-            if let Some(tx) = this.buffered.pop_front() {
-                this.push_tx(tx);
-            } else {
-                break
-            }
-        }
-
-        let mut stream_done = false;
-        loop {
-            match Stream::poll_next(Pin::new(&mut this.stream), cx) {
-                Poll::Ready(Some(tx)) => {
-                    if this.pending.len() < this.max_concurrent {
-                        this.push_tx(tx);
-                    } else {
-                        this.buffered.push_back(tx);
-                    }
-                }
-                Poll::Ready(None) => {
-                    stream_done = true;
-                    break
-                }
-                _ => break,
-            }
-        }
-
-        // poll running futures
-        if let tx @ Poll::Ready(Some(_)) = this.pending.poll_next_unpin(cx) {
-            return tx
-        }
-
-        if stream_done && this.pending.is_empty() {
-            // all done
-            return Poll::Ready(None)
-        }
-
-        Poll::Pending
     }
 }
 
@@ -342,7 +211,7 @@ mod tests {
             }))
             .await;
 
-        let stream = TransactionStream::new(
+        let stream = tx_stream::TransactionStream::new(
             &provider,
             stream::iter(txs.iter().cloned().map(|tx| tx.unwrap().transaction_hash)),
             10,
